Add optional LibPrice support for excluding items based on value
This currently limits items regardless of whether or not they're bound.
I'm not sure which is the right approach.  On the one hand, you can't
sell bound items.  On the other hand, it still indicates the item is
"valuable" and so you might want to keep it.

Note that dungeon items will be estimated at merchant price.