Fix handling of non-crafted craftable set items
Spellbreaker's Staff is a quest reward item which is not crafted,
but is part of the Shalidor's Curse craftable set.  Being a craftable
set, it cannot be reconstructed with transmute crystals.  This leads
to GetItemReconstructionCurrencyOptionCost returns nil, which crashes
when comparing nil <= 50.

Just exclude these items.  They might be easily reconstructed for a
high-level crafter, but other people may find it hard to recreate,
and want to use or transmute it.