Allow selection of crafted glyphs for levelling purposes
This adds a new option to allow marking of crafted glyphs for quick
levelling of enchanting.  Even if set, we restrict it to the following
cases:

1. The current character must not be max level in enchanting
2. Glyphs must be crafted by a different character

Self-crafted glyphs are likely to be intended for a different purpose,
such as daily crafting writs, master writs, actual gear enchanting,
selling, banking, mailing, and so on.  Also, if the current character
is already maxed in enchanting, the glyphs are likely intended for a
different character.